adding inception forward_materialized

+++ cuda/include/conv_cuda.h
@@ -13,7 +13,7 @@ void cudnn_mem_copy_gpu(int batch, int channels, int size, int padding, int stri
 
 void inc_max_pool_gpu(float* ptr_in_tensor, float* ptr_out_tensor, int in_size, int out_size, int channels, int batch, int padding, int stride, int k_size, int * ptr_location, int out_p_height, int out_p_width);
 
-void inc_add_gpu(float * ptr_in_tensor1, float * ptr_in_tensor2, float * ptr_out_tensor, float * ptr_location_tensor, int channels, int batch, int size, int p_height, int p_width, bool relu);
+void inc_add_gpu(float * ptr_in_tensor1, float * ptr_in_tensor2, float * ptr_out_tensor, int * ptr_location_tensor, int channels, int batch, int size, int p_height, int p_width, bool relu);
     
 void relu_fused_mem_copy_gpu(float *ptr_temp_tensor, float *ptr_out_tensor, float * ptr_biases, int * ptr_location, int batch, int p_height, int p_width, int channels, int size);
